tools/requirements.txt
azure-devops==7.1.0b4
msrest==0.7.1
requests==2.31.0
click==8.1.7
rich==13.7.1
python-dotenv==1.0.1
